% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradient descent.R
\name{grad_desc}
\alias{grad_desc}
\title{Title}
\usage{
grad_desc(Y, X, beta, step, n)
}
\arguments{
\item{Y}{Response vector}

\item{X}{Design matrix}

\item{beta}{Regression coefficients}

\item{step}{step size}
}
\value{

}
\description{
Title
}
\examples{
X <- matrix(c(1,1,1,1,2,2,2,2),nrow=4)
Y <- c(1,1,1,1)
n = 2
step = c(0.001,0.0005,0.01,0.02,0.03,0.04,0.05,0.1)
par(mfrow=c(2,2))
for(i in 1:length(step)){
  grad_desc(Y,X,beta,step[i],2)
}
}
